+++ cmd/clade/cmd/nextof_test.go
@@ -9,7 +9,7 @@ import (
 
 	"github.com/lesomnus/clade"
 	"github.com/lesomnus/clade/cmd/clade/cmd"
-	"github.com/lesomnus/clade/cmd/clade/cmd/internal/registry"
+	"github.com/lesomnus/clade/internal/registry"
 	"github.com/stretchr/testify/require"
 	"golang.org/x/exp/slices"
 )
