+++ internal/registry/errors.go
