+++ internal/registry/manifests.go
