+++ internal/registry/server.go
