+++ internal/registry/server_test.go
@@ -11,7 +11,7 @@ import (
 	"testing"
 
 	"github.com/distribution/distribution/v3/reference"
-	"github.com/lesomnus/clade/cmd/clade/cmd/internal/registry"
+	"github.com/lesomnus/clade/internal/registry"
 	"github.com/stretchr/testify/require"
 )
