+++ port.go
@@ -1,6 +1,8 @@
 package clade
 
 import (
+	"context"
+	"errors"
 	"fmt"
 	"os"
 	"path/filepath"
@@ -9,6 +11,7 @@ import (
 	"github.com/distribution/distribution/reference"
 	"github.com/lesomnus/clade/sv"
 	"github.com/lesomnus/pl"
+	"github.com/rs/zerolog"
 	"golang.org/x/exp/slices"
 	"gopkg.in/yaml.v3"
 )
@@ -168,4 +171,37 @@ func ReadPort(path string) (*Port, error) {
 	}
 
 	return port, nil
+}
+
+func ReadPortsFromFs(ctx context.Context, path string) ([]*Port, error) {
+	l := zerolog.Ctx(ctx)
+	l.Info().Str("path", path).Msg("read ports")
+
+	entries, err := os.ReadDir(path)
+	if err != nil {
+		return nil, fmt.Errorf("failed to read directory: %w", err)
+	}
+
+	ports := make([]*Port, 0)
+	for _, entry := range entries {
+		if !entry.IsDir() {
+			continue
+		}
+
+		port_path := filepath.Join(path, entry.Name(), "port.yaml")
+		l.Debug().Str("path", port_path).Msg("read port")
+
+		port, err := ReadPort(port_path)
+		if err != nil {
+			if errors.Is(err, os.ErrNotExist) {
+				continue
+			}
+
+			return nil, fmt.Errorf("failed to read port at %s: %w", port_path, err)
+		}
+
+		ports = append(ports, port)
+	}
+
+	return ports, nil
 }
