+++ util.go
@@ -1,8 +1,11 @@
 package clade
 
 import (
+	"fmt"
 	"os"
 	"path/filepath"
+
+	"github.com/lesomnus/pl"
 )
 
 // ResolvePath returns an absolute representation of joined path of `base` and `path`.
@@ -36,4 +39,33 @@ func ResolvePath(base string, path string, fallback string) (string, error) {
 	}
 
 	return filepath.Join(base, path), nil
+}
+
+func toString(value any) (string, bool) {
+	switch v := value.(type) {
+	case interface{ String() string }:
+		return v.String(), true
+	case string:
+		return v, true
+
+	default:
+		return "", false
+	}
+}
+
+func executeBeSingleString(executor *pl.Executor, pl *pl.Pl, data any) (string, error) {
+	results, err := executor.Execute(pl, data)
+	if err != nil {
+		return "", err
+	}
+	if len(results) != 1 {
+		return "", fmt.Errorf("expect result be sized 1 but was %d", len(results))
+	}
+
+	v, ok := toString(results[0])
+	if !ok {
+		return "", fmt.Errorf("expect result be string or stringer")
+	}
+
+	return v, nil
 }
